fix: chỉnh sửa quản lý db từ sequelize.sync() sang migration + seeder

+++ backend/src/models/courtPrice.js
@@ -66,7 +66,7 @@ const CourtPrice = sequelize.define(
       },
     },
     price: {
-      type: DataTypes.DOUBLE,
+      type: DataTypes.DECIMAL(10, 2),
       allowNull: false,
       validate: {
         notNull: { msg: "Price is required" },
